Move lambda_utils to legacy package
