+++ localstack/services/lambda_/legacy/lambda_utils.py
@@ -1,3 +1,4 @@
+"""Lambda utils for old Lambda provider"""
 import base64
 import json
 import logging
