+++ tests/aws/services/stepfunctions/v2/test_sfn_api.py
@@ -3,7 +3,7 @@ import json
 import pytest
 
 from localstack.aws.api.stepfunctions import StateMachineType
-from localstack.services.lambda_.lambda_utils import LAMBDA_RUNTIME_PYTHON39
+from localstack.services.lambda_.legacy.lambda_utils import LAMBDA_RUNTIME_PYTHON39
 from localstack.testing.pytest import markers
 from localstack.testing.snapshots.transformer import RegexTransformer
 from localstack.utils.strings import short_uid
